refactor(mcp): split 650-line mcp_server.py + roadmap + README
Maintainability (no behavior change — 42 mcp tests + full suite green):
- app/mcp/tools.py    — MCP_TOOLS schema (data only)
- app/mcp/handlers.py — _handle_* implementations + _installation_allowed + TOOL_HANDLERS
- app/mcp/mcp_server.py — transport/dispatch only (650 -> 105 lines); re-exports
  the public names via __all__ so existing imports/tests are unaffected.

Docs:
- docs/architecture/roadmap.md — forward-looking: prioritized improvements
  (real secret scanner, HA Redis, semantic memory, splitting remaining long
  files) + adoption plan + honest 'nothing fails/infected' framing.
- README: V6 changelog filled out, security-model no-RCE note, audit+roadmap links.

726 passed, 11 skipped, ruff clean.
